pyproject.toml
[tool.poetry]
name = "english-mingle-backend"
version = "0.1.0"
description = ""
authors = ["Donghyeok Lee <donghyukleee@gmail.com>"]
license = "MIT"
readme = "README.md"
packages = [{include = "english_mingle_backend"}]

[tool.poetry.dependencies]
python = "^3.11"
Django = "^4.2.3"
Pillow = "^10.0.0"
djangorestframework = "^3.14.0"
django-cors-headers = "^4.2.0"
pyjwt = "^2.8.0"
django-environ = "^0.10.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
